refactor(codex): consolidate native subagent relay
